Disable presence tracking on matrix.org
Presence is a feature that allows users to track whether the users of
another matrix system are online. This is fine, but it turns out it's
very CPU heavy. I am opting to disable this feature by default so that
users don't have their CPUs explode the moment they join a matrix.org
room.